// src/app/shared/services/api/model/get-blog-post-by-id-response.ts
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { GetAllCommentResponse } from './get-all-comment-response';


export interface GetBlogPostByIdResponse { 
    id?: number;
    title?: string;
    content?: string;
    authorName?: string;
    publishedDate?: string;
    comments?: Array<GetAllCommentResponse>;
    message?: string;
}